pyproject.toml
[project]
name = "conversation-ai"
version = "1.0.0"
description = "Multi-turn conversational AI chatbot using STT-LLM-TTS pipeline with LiveKit, Sarvam AI, and Google Gemini"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit>=1.0.0",
    "livekit-agents>=1.0.0",
    "livekit-plugins-silero>=1.0.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/conversation_ai"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
